Move BadTypeException to the correct namespace
